agregando layout

+++ frontend/src/context/CartContext.js
@@ -1,5 +1,7 @@
 import { createContext, useState, useEffect } from 'react'
 
+import Swal from 'sweetalert2';
+
 export const CartContext = createContext()
 
 const { Provider } = CartContext
@@ -19,14 +21,31 @@ export default function CartProvider({ children }) {
         setCart(arrayNuevo);
         setCantidadProductos(cantidadProductos + cantidad);
         localStorage.setItem("carrito", JSON.stringify(arrayNuevo));
+        Swal.fire({
+            position: 'center',
+            icon: 'success',
+            title: `${product.title} agregado`,
+            showConfirmButton: false,
+            timer: 1500
+        })
     }
     function delToCart(id) {
-        let carrito = cart.slice(0);
-        let carritoFinal = carrito.filter(e => e.id !== id)
-        let cantidadProductoEliminado = carrito.find(e => e.id === id).cantidad;
-        setCart(carritoFinal);
-        setCantidadProductos(cantidadProductos - cantidadProductoEliminado);
-        localStorage.setItem("carrito", JSON.stringify(carritoFinal));
+        Swal.fire({
+            position: 'center',
+            icon: 'warning',
+            title: `Desae eliminar el producto ${cart.find(e=>e.id===id).title}`,
+            showConfirmButton: true,
+            confirmButtonText:"Confirmar"
+        }).then((result) => {
+            if (result.isConfirmed) {
+                let carrito = cart.slice(0);
+                let carritoFinal = carrito.filter(e => e.id !== id)
+                let cantidadProductoEliminado = carrito.find(e => e.id === id).cantidad;
+                setCart(carritoFinal);
+                setCantidadProductos(cantidadProductos - cantidadProductoEliminado);
+                localStorage.setItem("carrito", JSON.stringify(carritoFinal));
+            }
+        })
     }
     function clearCart() {
         setCart([]);
